+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Features:
 Bugfixes:
 
 * Fixed an issue when plotting isotherms with and without secondary data simultaneously
+* Fixed error with magnitude of polarizability of adsorbate from database in microporous PSD
 
 
 1.1.0 (18-01-24)